+++ src/bin/platforms/linux.ts
@@ -138,27 +138,46 @@ export class LinuxInstaller {
         this.checkForRoot();
       }
 
-      const npmGlobalPath = child_process.execSync('/bin/echo -n "$(npm --no-update-notifier -g prefix)/lib/node_modules"').toString('utf8');
       const targetNodeVersion = child_process.execSync('node -v').toString('utf8').trim();
 
-      child_process.execSync('npm rebuild --unsafe-perm', {
-        cwd: process.env.UIX_BASE_PATH,
-        stdio: 'inherit',
-      });
+      if (this.isPackage() && process.env.UIX_USE_PNPM === '1' && process.env.UIX_CUSTOM_PLUGIN_PATH) {
+        // pnpm+package mode
+        const cwd = path.dirname(process.env.UIX_CUSTOM_PLUGIN_PATH);
 
-      if (all === true) {
-        // rebuild all modules
-        try {
-          child_process.execSync('npm rebuild --unsafe-perm', {
-            cwd: npmGlobalPath,
-            stdio: 'inherit',
-          });
-        } catch (e) {
-          this.hbService.logger('Could not rebuild all modules - check Homebridge logs.', 'warn');
+        if (!await fs.pathExists(cwd)) {
+          this.hbService.logger(`Path does not exist: "${cwd}"`, 'fail');
+          process.exit(1);
         }
-      }
 
-      this.hbService.logger(`Rebuilt modules in ${process.env.UIX_BASE_PATH} for Node.js ${targetNodeVersion}.`, 'succeed');
+        child_process.execSync(`pnpm -C "${cwd}" rebuild`, {
+          cwd: cwd,
+          stdio: 'inherit',
+        });
+        this.hbService.logger(`Rebuilt plugins in ${process.env.UIX_CUSTOM_PLUGIN_PATH} for Node.js ${targetNodeVersion}.`, 'succeed');
+      } else {
+        // normal global npm setups
+        const npmGlobalPath = child_process.execSync('/bin/echo -n "$(npm --no-update-notifier -g prefix)/lib/node_modules"').toString('utf8');
+
+        child_process.execSync('npm rebuild --unsafe-perm', {
+          cwd: process.env.UIX_BASE_PATH,
+          stdio: 'inherit',
+        });
+        this.hbService.logger(`Rebuilt homebridge-config-ui-x for Node.js ${targetNodeVersion}.`, 'succeed');
+
+        if (all === true) {
+          // rebuild all global node_modules
+          try {
+            child_process.execSync('npm rebuild --unsafe-perm', {
+              cwd: npmGlobalPath,
+              stdio: 'inherit',
+            });
+            this.hbService.logger(`Rebuilt plugins in ${npmGlobalPath} for Node.js ${targetNodeVersion}.`, 'succeed');
+          } catch (e) {
+            this.hbService.logger('Could not rebuild all plugins - check logs.', 'warn');
+          }
+        }
+
+      }
     } catch (e) {
       console.error(e.toString());
       this.hbService.logger('ERROR: Failed Operation', 'fail');